version 3.2.0

+++ samples/app/src/main/java/com/vk/sdk/sample/models/VKUser.kt
@@ -29,21 +29,21 @@ import android.os.Parcelable
 import org.json.JSONObject
 
 data class VKUser(
-        val id: Int = 0,
+        val id: Long = 0,
         val firstName: String = "",
         val lastName: String = "",
         val photo: String = "",
         val deactivated: Boolean = false) : Parcelable {
 
     constructor(parcel: Parcel) : this(
-            parcel.readInt(),
+            parcel.readLong(),
             parcel.readString()!!,
             parcel.readString()!!,
             parcel.readString()!!,
             parcel.readByte() != 0.toByte())
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeInt(id)
+        parcel.writeLong(id)
         parcel.writeString(firstName)
         parcel.writeString(lastName)
         parcel.writeString(photo)
@@ -64,7 +64,7 @@ data class VKUser(
         }
 
         fun parse(json: JSONObject)
-                = VKUser(id = json.optInt("id", 0),
+                = VKUser(id = json.optLong("id", 0),
                 firstName = json.optString("first_name", ""),
                 lastName = json.optString("last_name", ""),
                 photo = json.optString("photo_200", ""),
